Move calculateNextDecision into Rules
